// src/app/models/movie-upload.ts
export class MovieUpload {
    maPhim: number = 0;
    tenPhim: string = '';
    trailer: string = '';
    moTa: string = '';
    maNhom: string = 'GP01';
    ngayKhoiChieu: string = '';
    sapChieu: boolean = false;
    dangChieu: boolean = false;
    hot: boolean = true;
    danhGia: number = 0;
}